Amelioration de BuildingTopology

+++ src/warcraftPath/Patern1.java
@@ -15,7 +15,7 @@ public class Patern1 {
 	public static List<Position> pathconstruct() {
 		List<Position> path = new LinkedList<Position>();
 		for(int i=0; i<24; i++)
-			path.add(new Position(i * World.getSquareWidth() + World.getSquareWidth() / (float)2.0, 7 * World.getSquareHeight() + World.getSquareHeight() / (float)2.0));
+			path.add(new Position(i * World.getSquareWidth() + World.getSquareWidth() / (float)2.0,           (float)7 * (float)World.getSquareHeight() + (float)World.getSquareHeight() / (float)2.0));
 			
 		return path;
 	}
